Makefile: build geth and the plugin
I think this makes loading the plugin work.

The basic idea is as follows: in order to ensure that geth and the
plugin use exactly the same dependencies, the geth executable and plugin
must be built in the same Go module.

So we do just that. The Makefile builds a geth executable from the Go
module context of the plugin, and then builds the plugin itself using
the same go-ethereum dependency.